tests/krate/publish: Extract `manifest` module
